backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.6.0
xgboost==2.1.4
joblib==1.4.2
pdfplumber==0.11.5
SQLAlchemy==2.0.39
psycopg2-binary==2.9.10
python-jose==3.3.0
python-multipart==0.0.9
